Refactoring because screw using Docker for this.
