+++ chatbot.cfg
@@ -1,0 +1,7 @@
+[Main]
+HOST=irc.sithmail.com
+PORT=6697
+NICK=TheDoctor
+
+[Channels]
+chan1=#social
