--- tipbot/MANIFEST
@@ -1,3 +1,0 @@
-# file GENERATED by distutils, do NOT edit
-Legobot.py
-setup.py
