--- tipbot/config.fecn.template
@@ -1,13 +1,0 @@
-# List of global variables (all constants) to import.
-# format: <type>varname=value
-# types supported: int, str, bool, json
-# var names must be uppercase (will be cast to uppercase at load time)
-
-<str>HOST=irc.sithmail.com
-<str>HOSTPW= 
-<int>PORT=6697
-<str>NICK = TheDoctor
-<str>ROOM = #social
-<str>ROOMPW = 
-<int>SNARKHOLDDOWN = 15
-<bool>ISSSL = True
