--- tipbot/snark.fecn
@@ -1,80 +1,0 @@
-#list of snarky comments for MrQ to read
-I wish sarcasm was available as a font
-I'm allergic to stupidity so I'll break out in sarcasm
-The thing about quotes on the internet is that you cannot confirm their validity - Abraham Lincoln
-Don't judge me because I'm quiet. Nobody plans a murder out loud
-I'm somewhat of a BSer myself, but occasionally I like to listen to a real pro. Please go on!
-When they discover the center of the universe, you will be disappointed to find that it isn't you.
-Some people can't sleep because they have insomnia. I can't sleep because I have an internet connection.
-As I have grown older, I've learned that pleasing everyone is impossible, but pissing everyone off is a piece of cake
-Careful not to trip over my amazingness... It's everywhere
-I'm sorry to be the one to have to tell you this: you have one of the most severe cases of terminal stupidity I have ever diagnosed. There is no cure
-Bro, I'm going to need you to put those few remaining brain cells together and work with me here, ok?
-Today was going great, until... you
-Everyone brings joy to this team... some when entering, some when leaving
-Rasin cookies that look like chocolate chip cookies are the main reason why I have trust issues
-I may seem calm, but in my mind I have killed you three times
-I am proficient in three languages: python, sarcasm, and profanity
-An apple a day will keep anyone away if you throw it hard enough
-Shh! I'm hiding from stupid people!
-It isn't that I'm not a people person, it's that I'm not a stupid people person
-My name is Inigo Montoya, you ignored my collab request, prepare to die
-If Frodo can get the ring to Mordor, you can solve that case
-I used to care, but now I take a pill for that
-The difference between genius and stupidity is that genius has limits
-If you browse imgur long enough, that case will solve itself
-Arguing with an Pavan is like wrestling with a pig in mud. After about 2 hours, you realize he likes it
-Sometimes when the internet is down I forget that the rest of my computer still works
-On the internet you can be anything that you want, it's so strange that you choose to be stupid
-I find it incredibly demeaning that I have to share the internet with the likes of you... and Internet Explorer
-Do you ever look at your life and say 'what has the internet done to me?'
-Money doesn't buy happiness, but it pays for the internet... which is almost the same thing
-With that astronomically intellectual statement, I would have though you worked for VCE
-You are why we can't have nice things
-Did you try to open a .part file again?
-Dude, even Microsoft could have figured this out.
-Even Briggs knows that doesn't work. Serious?
-I don't always write C but wheSEGMENTATION FAULT CORE DUMPED
-Some people, when confronted with a problem, think, "I know, I'll use threads," and then two they hav erpoblesms.
-http://i.imgur.com/6oW6Qhv.gif
-http://i.imgur.com/GPqtwr7.gif
-http://i.imgur.com/ElCPi.gif
-http://i.imgur.com/R2v3c6H.gif
-http://i.imgur.com/ZQ4RyHo.gif
-http://i.imgur.com/LCUjeU1.gif
-http://i.imgur.com/woZL9TO.gif
-http://i.imgur.com/YkoCx.gif
-http://i.imgur.com/RVBi3T1.gif
-http://i.imgur.com/PeOkjoB.gif
-http://i.imgur.com/G4QbkbX.gif
-http://i.imgur.com/0U2cnck.gif
-http://i.imgur.com/6uasde9.gif
-http://i.imgur.com/l9zKYHF.gif
-http://i.imgur.com/Lmbhspn.gif
-http://i.imgur.com/BVzQk.gif
-http://i.imgur.com/3Usiw.gif
-http://i.imgur.com/whzga.gif
-http://i.imgur.com/CaTwx.gif
-http://i.imgur.com/tw7aQ.gif
-http://i.imgur.com/HpvWY.gif
-http://i.imgur.com/dyeHb.gif
-http://i.imgur.com/2W7lX.gif
-http://i.imgur.com/nc9j0.gif
-http://i.imgur.com/9Xy2Y.gif
-http://i.imgur.com/qnniC.gif
-http://i.imgur.com/HKyii.gif
-http://i.imgur.com/He43V.gif
-http://i.imgur.com/RERnH.gif
-http://i.imgur.com/TghktaI.gif
-http://i.imgur.com/zS5rRxa.gif
-http://i.imgur.com/uvz4Qys.gif
-http://i.imgur.com/sYquW0y.gif
-http://i.imgur.com/Dg5J7.gif
-http://i.imgur.com/PDxeZ.gif
-http://i.imgur.com/EkvDRTI.gif
-http://i.imgur.com/SbISy.gif
-http://i.imgur.com/VSYRt.gif
-http://i.imgur.com/ET0qANB.gif
-http://i.imgur.com/lgF6RhJ.gif
-http://i.imgur.com/fEFRJqV.gif
-http://i.imgur.com/m2r5gtd.gif
